feat(api-headless-cms-ddb): get value on given object or array path
