+++ packages/api-headless-cms-ddb/src/operations/entry/filtering/index.ts
@@ -1,0 +1,2 @@
+export { filter } from "./filter";
+export { sort } from "./sort";
